Project_spring_2/requirements.txt
aiogram==3.4.1
aiohttp==3.9.5
python-dotenv==1.0.1
openai